nimaprojectnew/requirements.txt
Django==1.11.10
django-filter==1.1.0
djangorestframework==3.7.7
Markdown==2.6.11
Pillow==5.0.0
pkg-resources==0.0.0
pytz==2018.3
qrcode==5.3
reportlab==3.4.0
six==1.11.0
